Move `KeychainUtils` to WordPressShared
