+++ WordPress/WordPressTest/SharedDataIssueSolverTests.swift
@@ -1,5 +1,6 @@
 import XCTest
 @testable import WordPress
+@testable import WordPressShared
 
 class SharedDataIssueSolverTests: XCTestCase {
 
@@ -143,46 +144,6 @@ private final class CoreDataStackMock: CoreDataStack {
 
     func performAndSave(_ aBlock: @escaping (NSManagedObjectContext) -> Void) {}
     func performAndSave(_ aBlock: @escaping (NSManagedObjectContext) -> Void, completion: (() -> Void)?, on queue: DispatchQueue) {}
-}
-
-// MARK: - KeychainUtilsMock
-
-private final class KeychainUtilsMock: KeychainUtils {
-
-    var sourceAccessGroup: String?
-    var destinationAccessGroup: String?
-    var shouldThrowError = false
-    var passwordToReturn: String? = nil
-    var storeShouldThrow = false
-    var storedPassword: String? = nil
-    var storedUsername: String? = nil
-    var storedServiceName: String? = nil
-    var storedAccessGroup: String? = nil
-
-    override func copyKeychain(from sourceAccessGroup: String?, to destinationAccessGroup: String?, updateExisting: Bool = true) throws {
-        if shouldThrowError {
-            throw NSError(domain: "", code: 0)
-        }
-
-        self.sourceAccessGroup = sourceAccessGroup
-        self.destinationAccessGroup = destinationAccessGroup
-    }
-
-    override func password(for username: String, serviceName: String, accessGroup: String? = nil) throws -> String? {
-        return passwordToReturn
-    }
-
-    override func store(username: String, password: String, serviceName: String, accessGroup: String? = nil, updateExisting: Bool) throws {
-        if storeShouldThrow {
-            throw NSError(domain: "", code: 0)
-        }
-
-        storedUsername = username
-        storedPassword = password
-        storedServiceName = serviceName
-        storedAccessGroup = accessGroup
-    }
-
 }
 
 // MARK: - Mock Local File Store
